requirements.txt
Django==2.2.4
psycopg2==2.8.3
gunicorn==19.9.0
djangorestframework==3.12.4
django-cors-headers==3.7.0